Refactor tuning/testers into tuning directory
