+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/tuning/motor/LaunchMotorTuner.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.tuning.PIDF;
+package org.firstinspires.ftc.teamcode.tuning.motor;
 
 import android.util.Log;
 
@@ -11,18 +11,11 @@ import com.seattlesolvers.solverslib.command.CommandOpMode;
 import com.seattlesolvers.solverslib.controller.PIDFController;
 import com.seattlesolvers.solverslib.util.TelemetryData;
 
-import org.firstinspires.ftc.teamcode.globals.Constants;
 import org.firstinspires.ftc.teamcode.globals.Robot;
 
 @Config
 @TeleOp(name = "LaunchMotorTuner")
 public class LaunchMotorTuner extends CommandOpMode {
-    // D, 0.000_
-    // F, 0.000_
-    // I, 0
-    // maxPowerConstant, 1.0
-    // P, 0.00_
-
     public static double P = 0.00;
     public static double I = 0;
     public static double D = 0.000;
@@ -39,8 +32,6 @@ public class LaunchMotorTuner extends CommandOpMode {
 
     @Override
     public void initialize() {
-        // Must have for all opModes
-        Constants.OP_MODE_TYPE = Constants.OpModeType.TELEOP;
         launcherPIDF.setTolerance(POS_TOLERANCE, 0);
 
         // Resets the command scheduler
@@ -52,20 +43,16 @@ public class LaunchMotorTuner extends CommandOpMode {
 
     @Override
     public void run() {
-        // Keep all the has movement init for until when TeleOp starts
-        // This is like the init but when the program is actually started
         if (timer == null) {
             timer = new ElapsedTime();
         }
 
         double motorVel = robot.launchMotors.getVelocity();
 
-        launcherPIDF.setPIDF(P, I, D, F);
-
-        launcherPIDF.setTolerance(POS_TOLERANCE, 0);
-
         TARGET_VEL = Math.max(0, TARGET_VEL);
 
+        launcherPIDF.setPIDF(P, I, D, F);
+        launcherPIDF.setTolerance(POS_TOLERANCE, 0);
         launcherPIDF.setSetPoint(TARGET_VEL);
 
         double power = launcherPIDF.calculate(motorVel, TARGET_VEL);
