+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/tuning/servo/LaunchServosTuner.java
@@ -17,7 +17,7 @@ import org.firstinspires.ftc.teamcode.globals.Robot;
 
 @Config
 @TeleOp
-public class LaunchServosTuner extends CommandOpMode {
+public class     LaunchServosTuner extends CommandOpMode {
     public GamepadEx driver;
     public GamepadEx operator;
