--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/tuning/servo/sample/DoubleServoTuner.java
@@ -1,84 +1,0 @@
-package org.firstinspires.ftc.teamcode.tuning.servo.sample;
-
-import com.acmerobotics.dashboard.FtcDashboard;
-import com.acmerobotics.dashboard.config.Config;
-import com.acmerobotics.dashboard.telemetry.MultipleTelemetry;
-import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
-import com.qualcomm.robotcore.util.ElapsedTime;
-import com.seattlesolvers.solverslib.command.CommandOpMode;
-import com.seattlesolvers.solverslib.command.InstantCommand;
-import com.seattlesolvers.solverslib.gamepad.GamepadEx;
-import com.seattlesolvers.solverslib.gamepad.GamepadKeys;
-import com.seattlesolvers.solverslib.geometry.Pose2d;
-import com.seattlesolvers.solverslib.util.TelemetryData;
-
-import org.firstinspires.ftc.teamcode.globals.Constants;
-import org.firstinspires.ftc.teamcode.globals.Robot;
-
-//@Config
-//@TeleOp
-public class DoubleServoTuner extends CommandOpMode {
-    public GamepadEx driver;
-    public GamepadEx operator;
-
-    public ElapsedTime timer;
-
-    public static double LEFT_SERVO_POS = 0.0;
-    public static double RIGHT_SERVO_POS = 0.0;
-    public static double BOTH_SERVO_POS = 0.0;
-
-    public static boolean MOVE_BOTH = false;
-
-    TelemetryData telemetryData = new TelemetryData(new MultipleTelemetry(telemetry, FtcDashboard.getInstance().getTelemetry()));
-
-    private final Robot robot = Robot.getInstance();
-
-    @Override
-    public void initialize() {
-        // Must have for all opModes
-        Constants.OP_MODE_TYPE = Constants.OpModeType.TELEOP;
-
-        // Resets the command scheduler
-        super.reset();
-
-        // Initialize the robot (which also registers subsystems, configures CommandScheduler, etc.)
-        robot.init(hardwareMap);
-
-        driver = new GamepadEx(gamepad1);
-        operator = new GamepadEx(gamepad2);
-
-        // Driver controls
-        // Reset heading
-        driver.getGamepadButton(GamepadKeys.Button.DPAD_UP).whenPressed(
-                new InstantCommand(() -> robot.drive.setPose(new Pose2d()))
-        );
-    }
-
-    @Override
-    public void run() {
-        // Keep all the has movement init for until when TeleOp starts
-        // This is like the init but when the program is actually started
-        if (timer == null) {
-            robot.initHasMovement();
-            timer = new ElapsedTime();
-        }
-
-        if (MOVE_BOTH) {
-            robot.leftIntakePivot.set(BOTH_SERVO_POS);
-            robot.rightIntakePivot.set(BOTH_SERVO_POS);
-        } else {
-            robot.leftIntakePivot.set(LEFT_SERVO_POS);
-            robot.rightIntakePivot.set(RIGHT_SERVO_POS);
-        }
-
-        telemetryData.addData("Loop Time", timer.milliseconds());
-        timer.reset();
-
-        telemetryData.addData("LEFT_SERVO_POS", LEFT_SERVO_POS);
-        telemetryData.addData("RIGHT_SERVO_POS", RIGHT_SERVO_POS);
-
-        // DO NOT REMOVE ANY LINES BELOW! Runs the command scheduler and updates telemetry
-        super.run();
-        telemetryData.update();
-    }
-}
